PrimaryInstance initializes translation state, implements EvaluationContext
